feat(auth): implement logout functionality with JWT blacklisting and error handling; update Claims struct feat(user): add delete_user_by_id function and update user routes refactor: change token duration and timestamp types to u64 for consistency

+++ src/utils/auth.rs
@@ -7,14 +7,14 @@ use jsonwebtoken::{DecodingKey, EncodingKey, Header, Validation, decode, encode}
 
 use crate::models::auth::{Claims, UserRole};
 
-const DURATION: usize = 3600;
+const DURATION: u64 = 3600;
 
 pub fn genrate_jwt(
     role: UserRole,
     user_id: Uuid,
     key: EncodingKey,
 ) -> Result<String, jsonwebtoken::errors::Error> {
-    let now = chrono::Utc::now().timestamp() as usize;
+    let now = chrono::Utc::now().timestamp() as u64;
     let jti = Uuid::new_v4();
     let claims = Claims {
         sub: user_id,
